don't allow language tokens in output for 1:many decoding
Summary:
Because of the way language ID tokens were introduced for training 1:N MT models, we sometimes see the artifact of the model producing language tokens in the output (see T119348697 for details). This change prevents that by simply zeroing out their probabilities during beam search.

Edit: it turns out that the majority of cases of target language token appearing in the output are due to UNK replacement, where the language ID in source prefix was the "most-attended-to" source token at the step where the UNK was produced. Thus, this change zeros out the attention weights for prefix tokens in the source sequence.

Reviewed By: theweiho

Differential Revision: D36282175

fbshipit-source-id: fb7b2bfd3a8c1c66563ea509e68ab742a831ba4a